Device interface and PCI bus stub
