+++ sys/devices.c
@@ -1,0 +1,16 @@
+#include <drivers.h>
+#include <common.h>
+#include <stdc.h>
+
+int bus_register(bus_type_t *bus) {
+  log("Registering a bus");
+  return 0;
+}
+int device_register(device_t *device) {
+  log("Registering a device");
+  return 0;
+}
+int driver_register(driver_t *driver) {
+  log("Registering a driver");
+  return 0;
+}
